skyblockparser-main/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "skyblockparser"
version = "2.0"
authors = [
  { name="nom", email="noemtdev@gmail.com" },
]
description = "A package allowing you to parse skyblock profile information including networth."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords=['python', 'hypixel', 'skyblock', 'hypixel skyblock']
dependencies = ['aiohttp', 'pillow']

[project.urls]
"Homepage" = "https://github.com/noemtdev/skyblockparser"
"Bug Tracker" = "https://github.com/noemtdev/skyblockparser/issues"